requirements.txt
kubernetes==25.3.0
kafka-python
prometheus-client
pandas
numpy
